% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_add_method.R
\name{add,PopulationPyramid,PopulationPyramid-method}
\alias{add,PopulationPyramid,PopulationPyramid-method}
\title{Add Two PopulationPyramid Objects Together}
\usage{
\S4method{add}{PopulationPyramid,PopulationPyramid}(obj1, obj2)
}
\arguments{
\item{obj1}{\code{PopulationPyramid} object}

\item{obj2}{\code{PopulationPyramid} object}
}
\value{
Updated \code{PopulationPyramid} object
}
\description{
Add Two PopulationPyramid Objects Together
}
