% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_add_method.R
\name{add,TaskParameters,TaskParameters-method}
\alias{add,TaskParameters,TaskParameters-method}
\title{Add Two TaskParameters Objects Together}
\usage{
\S4method{add}{TaskParameters,TaskParameters}(obj1, obj2)
}
\arguments{
\item{obj1}{\code{TaskParameters} object}

\item{obj2}{\code{TaskParameters} object}
}
\value{
Updated \code{TaskParameters} object
}
\description{
Add Two TaskParameters Objects Together
}
