% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_exp_control.R
\name{addRatesMatricesToPopulationChangeRates}
\alias{addRatesMatricesToPopulationChangeRates}
\title{Compute Time-Series of Population Rate Parameters}
\usage{
addRatesMatricesToPopulationChangeRates(pcr, years, stochasticParms = NULL)
}
\arguments{
\item{pcr}{Population Change Rates structure}

\item{years}{Range of years to compute}

\item{stochasticParms}{Parameters controlling stochastic variation of generated rates}
}
\value{
Population Change Rates structure with predicted rates matrices added
}
\description{
Compute Time-Series of Population Rate Parameters
}
\examples{
\dontrun{
initPop <- pacehrh:::loadInitialPopulation(sheetName = "Flat_Population")
pcr <- pacehrh:::loadPopulationChangeRates(sheetName = "Flat_Rates")
pars <- pacehrh:::loadStochasticParameters(sheetName = "Flat_StochasticParms")
# Turn off stochasticity and generate several years of rates
years <- 2020:2040
pcr <- pacehrh:::addRatesMatricesToPopulationChangeRates(pcr, years, NULL)
}
}
