% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_lint.R
\name{CheckInputExcelFileFormat}
\alias{CheckInputExcelFileFormat}
\title{Perform Sanity Checks On Input Excel File}
\usage{
CheckInputExcelFileFormat(
  inputFile = NULL,
  outputFile = NULL,
  scenarioSheet = "Scenarios",
  seasonalityOffsetsSheet = "SeasonalityOffsets",
  noDate = FALSE
)
}
\arguments{
\item{inputFile}{Excel file to examine. If NULL, check the file defined in the global Configuration.}

\item{outputFile}{Results output file. If NULL, results are printed to the console.}

\item{scenarioSheet}{Name of the sheet with scenario details. Default = "Scenarios".}

\item{seasonalityOffsetsSheet}{Name of the sheet with per-task seasonality offset details. Default = "SeasonalityOffsets"}

\item{noDate}{Suppress date stamp from output. Default = FALSE.}
}
\value{
Error code.
0 = Success
< 0 (negative) = Fatal error. Nothing will work right.
> 0 (positive) = Warning. Some scenarios might fail, some might succeed.
}
\description{
Do a set of sanity checks on the input Excel file.
}
\examples{
\dontrun{
errCode <- pacehrh::CheckInputExcelFileFormat(outputFile = "config_file_check.txt")
}
}
