% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_population_utils.R
\name{ComputeApplicablePopulationMatrices}
\alias{ComputeApplicablePopulationMatrices}
\title{Generate Matrices of Applicable Populations}
\usage{
ComputeApplicablePopulationMatrices(
  results,
  labels = GPE$populationLabels$Labels
)
}
\arguments{
\item{results}{Results structure as returned by \code{RunExperiments()}}
}
\value{
matrix
}
\description{
Generate Matrices of Applicable Populations
}
\examples{
\dontrun{
# This code won't be run
}
}
