% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_population_utils.R
\name{ComputeApplicablePopulationMatrix}
\alias{ComputeApplicablePopulationMatrix}
\title{Generate Matrix of Applicable Populations}
\usage{
ComputeApplicablePopulationMatrix(result, labels = GPE$populationLabels$Labels)
}
\arguments{
\item{result}{One entry from the results structure as returned by \code{RunExperiments()}}
}
\value{
An applicable population matrix (labels x years)
}
\description{
Generate Matrix of Applicable Populations
}
\examples{
\dontrun{
}
}
