% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_post_process.R
\name{ComputeSummaryStats}
\alias{ComputeSummaryStats}
\title{Compute Summary Statistics}
\usage{
ComputeSummaryStats(DR = NULL, CA = NULL)
}
\arguments{
\item{DR}{From \code{ReadAndCollateSuiteResults()} function}

\item{CA}{From \code{ComputeCadreAllocations()} function}
}
\value{
List of computed summary tables
}
\description{
Compute Summary Statistics
}
\examples{
\dontrun{
resultsFiles <- c("results/results_BasicModel_Jul04.csv",
                  "results/results_ComprehensiveModel_Jul04.csv",
                  "results/results_MergedModel_Jul04.csv")

DR <- pacehrh::ReadAndCollateSuiteResults(files = resultsFiles)
CA <- pacehrh::ComputeCadreAllocations(DR)
stats <- pacehrh::ComputeSummaryStats(DR,CA)
}
}
