% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_exp_control.R
\name{ConfigureExperimentValues}
\alias{ConfigureExperimentValues}
\title{Generate A New Set Of Stochastic Variations}
\usage{
ConfigureExperimentValues()
}
\value{
NULL (invisible)
}
\description{
Create the configuration values for an individual model experiment.
}
\examples{
\dontrun{
library(pacehrh)
pacehrh::Trace(TRUE)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

set.seed(54321)

scenarioData <- SaveBaseSettings(scenario)
ConfigureExperimentValues()
results <- RunExperiment()
}
}
