% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_scenarios_manage.R
\name{CreateScenariosTable}
\alias{CreateScenariosTable}
\title{Create A New Scenarios Table}
\usage{
CreateScenariosTable()
}
\value{
Blank scenarios table
}
\description{
Create A New Scenarios Table
}
\examples{
\dontrun{
tbl <- pacehrh::CreateScenariosTable()
tbl <- pacehrh::AddScenario(tbl, "Scenario 1")
}
}
