% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_scenarios_manage.R
\name{DeleteScenario}
\alias{DeleteScenario}
\title{Delete A Scenario From A Scenarios Table}
\usage{
DeleteScenario(tbl, UniqueID)
}
\arguments{
\item{tbl}{Scenarios table}

\item{UniqueID}{Scenario name. Must be unique.}
}
\value{
Updated scenarios table
}
\description{
Delete A Scenario From A Scenarios Table
}
\examples{
\dontrun{
tbl <- pacehrh::CreateScenariosTable()
tbl <- pacehrh::AddScenario(tbl, "Scenario 1")
tbl <- pacehrh::AddScenario(tbl, "Scenario 2")
tbl <- pacehrh::AddScenario(tbl, "Scenario 3")
tbl <- pacehrh::DeleteScenario(tbl, "Scenario 2")
}
}
