% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_experiment.R
\name{.computePopulationRangeMatrices}
\alias{.computePopulationRangeMatrices}
\title{Compute Population Range Sizes Based On Population Predictions}
\usage{
.computePopulationRangeMatrices(populations, popRanges)
}
\arguments{
\item{populations}{Population predictions as returned by [ComputePopulationProjection()]}

\item{popRanges}{Population range definitions as configured by [InitializePopulation()]}
}
\value{
List of population range sizes
}
\description{
Compute Population Range Sizes Based On Population Predictions
}
\examples{
\dontrun{
EXP$populationRangeMatrices <-
  .computePopulationRangeMatrices(EXP$demographics, BVE$populationRangesTable)
}
}
