% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_prevalence_matrix.R
\name{generatePrevalenceRatesMatrix}
\alias{generatePrevalenceRatesMatrix}
\title{Generate A Matrix Of Stochastically Varied Prevalence/Incidence Rates}
\usage{
generatePrevalenceRatesMatrix()
}
\value{
Rates matrix
}
\description{
Generate A Matrix Of Stochastically Varied Prevalence/Incidence Rates
}
