% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_rates_matrix.R
\name{generateRatesMatrix}
\alias{generateRatesMatrix}
\title{Generate A Matrix Of Stochastically Varied Population Change Rates}
\usage{
generateRatesMatrix(label = NULL)
}
\arguments{
\item{label}{Lookup string for the rates type}
}
\value{
Rates matrix
}
\description{
Generate A Matrix Of Stochastically Varied Population Change Rates
}
