% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_utils.R
\name{GetSuiteRates}
\alias{GetSuiteRates}
\title{Generate A Long-Format Data Frame Of Suite Mortality and Fertility Rates}
\usage{
GetSuiteRates(results, rateCategory = .populationRateCategories[1])
}
\arguments{
\item{results}{Results structure (as returned by \code{\link[=RunExperiments]{RunExperiments()}})}

\item{rateCategory}{Rate type ({"femaleFertility" | "maleFertility" | "femaleMortality" | "maleMortality"})}
}
\value{
Data frame
}
\description{
Generate A Long-Format Data Frame Of Suite Mortality and Fertility Rates
}
\examples{
\dontrun{
}
}
