% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_getValue_method.R
\name{getValue,PopulationChangeParameters,character-method}
\alias{getValue,PopulationChangeParameters,character-method}
\title{Get Population Change Parameter Values}
\usage{
\S4method{getValue}{PopulationChangeParameters,character}(object, label)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}

\item{label}{Name of variable to read. Allowed variable names are \code{
"FertRate",
"FertYears",
"AnnualBirthRateAll",
"AnnualBirthRate15_19",
"AnnualBirthRate20_29",
"AnnualBirthRate30_39",
"AnnualBirthRate40_49",
"MortalityInfants",
"Mortality1_4",
"Mortality5_9",
"Mortality10_14",
"Mortality15_19",
"Mortality20_24",
"MortalityAdultF",
"MortalityAdultM",
"AnnualBirthRate20_24",
"AnnualBirthRate25_29",
"AnnualBirthRate30_34",
"AnnualBirthRate35_39",
"AnnualBirthRate40_44",
"AnnualBirthRate45_49"
}}
}
\value{
Variable value
}
\description{
Get Population Change Parameter Values
}
