% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_cadre_config.R
\name{InitializeCadreRoles}
\alias{InitializeCadreRoles}
\title{Initialize Cadre Roles}
\usage{
InitializeCadreRoles(...)
}
\arguments{
\item{...}{Parameters passed through to loadCadreRoles()}
}
\value{
NULL (invisible)
}
\description{
Read the cadre roles information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()
pacehrh::InitializeCadreRoles()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(
    scenarioName = scenario,
    trials = 100
  )
}
}
