% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_task_config.R
\name{InitializeHealthcareTasks}
\alias{InitializeHealthcareTasks}
\title{Initialize Healthcare Task Information}
\usage{
InitializeHealthcareTasks(...)
}
\arguments{
\item{...}{Parameters passed through to \code{\link[=loadTaskParameters]{loadTaskParameters()}}}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare task information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeHealthcareTasks()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
