% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_population_config.R
\name{InitializePopulation}
\alias{InitializePopulation}
\title{Initialize Population Data}
\usage{
InitializePopulation()
}
\value{
NULL (invisible)
}
\description{
Load basic population information into the global package environment,
from which it can be used for later processing.
}
\examples{
\dontrun{
ehep::InitializePopulation()
}
}
