% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_scenarios_config.R
\name{InitializeScenarios}
\alias{InitializeScenarios}
\title{Initialize Experiment Scenario Information}
\usage{
InitializeScenarios(loadFromExcel = TRUE, ...)
}
\arguments{
\item{loadFromExcel}{If TRUE, initialize the scenarios list from the model
inputs Excel file. If FALSE, initialize with a blank scenarios table.}

\item{...}{See \code{\link[=loadScenarios]{loadScenarios()}}}
}
\value{
NULL (invisible)
}
\description{
Read experiment scenario information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
