% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_stochasticity_config.R
\name{InitializeStochasticParameters}
\alias{InitializeStochasticParameters}
\title{Initialize Stochastic Parameters}
\usage{
InitializeStochasticParameters(...)
}
\arguments{
\item{...}{Parameters passed through to \code{\link[=loadStochasticParameters]{loadStochasticParameters()}}}
}
\value{
NULL (invisible)
}
\description{
Read the stochasticity information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeHealthcareTasks()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
