% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_cadre_config.R
\name{loadCadreRoles}
\alias{loadCadreRoles}
\title{Load Cadre Roles}
\usage{
loadCadreRoles(sheetName = .defaultCadreRolesSheet)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
Dataframe of cadre roles information
}
\description{
Load Cadre Roles
}
