% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_coverage_config.R
\name{loadCoverageRates}
\alias{loadCoverageRates}
\title{Load CoverageRates Values}
\usage{
loadCoverageRates(sheetName = .defaultCoverageRatesSheet)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
Data table of per-year task coverage levels
}
\description{
Read the annual task coverage values from the model inputs Excel file. The
name and location of the model inputs Excel file is loaded from the global
configuration JSON file.
}
