% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_config.R
\name{loadGlobalConfig}
\alias{loadGlobalConfig}
\title{Load Global Configuration}
\usage{
loadGlobalConfig(path = "./globalconfig.json")
}
\arguments{
\item{path}{Location of global configuration file}
}
\value{
NULL (invisible)
}
\description{
Finds and loads global configuration data from a JSON file, including
the location of the input Excel data file.
}
