% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_population_config.R
\name{loadInitialPopulation}
\alias{loadInitialPopulation}
\title{Load Initial Population}
\usage{
loadInitialPopulation(sheetName = .defaultPopSheet)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
Tibble with three population pyramid fields:
\code{Female}, \code{Male} and \code{Total}
}
\description{
Read the initial population pyramid from the model inputs Excel file. The
name and location of the model inputs Excel file is loaded from the global
configuration JSON file.
}
