% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_scenarios_config.R
\name{loadScenarios}
\alias{loadScenarios}
\title{Load Experiment Scenario Information}
\usage{
loadScenarios(sheetName = "Scenarios")
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
Data frame of experiment scenario parameters
}
\description{
Read the experiment scenario information from the model inputs Excel file.
The name and location of the model inputs Excel file is loaded from the
global configuration JSON file.
}
