% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_seasonality_config.R
\name{loadSeasonalityCurves}
\alias{loadSeasonalityCurves}
\title{Load Seasonality Curves}
\usage{
loadSeasonalityCurves(sheetNameCurves = .defaultSeasonalityCurvesSheet)
}
\arguments{
\item{sheetNameCurves}{Sheet name from the model input Excel file}
}
\value{
Seasonality curves data frame.
}
\description{
Read the seasonality curves from the model inputs Excel file. The
name and location of the model inputs Excel file is loaded from the global
configuration JSON file.
}
