% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_stochasticity_config.R
\name{loadStochasticParameters}
\alias{loadStochasticParameters}
\title{Load Stochastic Parameters}
\usage{
loadStochasticParameters(sheetName = .defaultStochasticParametersSheet)
}
\arguments{
\item{sheetName}{Sheet name from the model input Excel file}
}
\value{
Dataframe of stochastic parameters
}
\description{
Load Stochastic Parameters
}
