% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_initialize.R
\name{PaceInitialize}
\alias{PaceInitialize}
\title{Initialize PACE-HRH}
\usage{
PaceInitialize(
  globalConfigFile = NULL,
  inputFile = NULL,
  populationSheet = NULL,
  forceGlobalConfigReload = TRUE
)
}
\arguments{
\item{globalConfigFile}{stuff stuff stuff}

\item{inputFile}{One entry from the results structure as returned by
\code{RunExperiments()}}

\item{populationSheet}{stuff stuff stuff}

\item{forceGlobalConfigReload}{TRUE/FALSE, default = TRUE. Force a reload of the global configuration if `globalConfigFile` is specified.}
}
\value{
TRUE/FALSE
}
\description{
Initialize PACE-HRH
}
\examples{
\dontrun{
}
}
