% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot_fertility.R
\name{PlotFertilityRatesStats}
\alias{PlotFertilityRatesStats}
\title{Plot Fertility Rates Statistics}
\usage{
PlotFertilityRatesStats(results, se = FALSE, type = "lines", log = TRUE)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}

\item{se}{Whether to show standard error or standard deviation confidence intervals.
Note: the \code{se} parameter does not apply to the "boxplot" type.}

\item{type}{Plot type (options = "lines", "ribbon", "boxplot")}

\item{log}{Use log scale for rates (default = TRUE). Applies only to "boxplot" type.}
}
\value{
ggplot grob
}
\description{
Plot Fertility Rates Statistics
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)

g <- ehep::PlotFertilityRatesStats(results, type = "boxplot", log = FALSE)
print(g)

g <- ehep::PlotFertilityRatesStats(results, type = "boxplot", log = TRUE)
print(g)

g <- ehep::PlotFertilityRatesStats(results, se = FALSE, type = "lines")
print(g)

g <- ehep::PlotFertilityRatesStats(results, se = TRUE, type = "lines")
print(g)

g <- ehep::PlotFertilityRatesStats(results, se = FALSE, type = "ribbon")
print(g)

g <- ehep::PlotFertilityRatesStats(results, se = TRUE, type = "ribbon")
print(g)
}
}
