% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_plot.R
\name{PlotPopulationCurves}
\alias{PlotPopulationCurves}
\title{Plot A Family Of Population Curves}
\usage{
PlotPopulationCurves(..., xaxis = NULL, colors = NULL, shapes = NULL)
}
\arguments{
\item{...}{Population vectors}

\item{xaxis}{Vector of age values for the x axis}

\item{colors}{Vector of colors for each population line}

\item{shapes}{Vector of shapes for each population point}
}
\value{
ggplot grob
}
\description{
Plot A Family Of Population Curves
}
\examples{
\dontrun{
g <- pacehrh::PlotPopulationCurves(
  results[[26]]$Population[['2020']]$Male,
  results[[26]]$Population[['2021']]$Male,
  results[[26]]$Population[['2022']]$Male,
  results[[26]]$Population[['2023']]$Male,
  results[[26]]$Population[['2024']]$Male,
  results[[26]]$Population[['2025']]$Male,
  xaxis = pacehrh:::GPE$ages,
  colors = c("royalblue4", "royalblue3", "royalblue2", "royalblue1"),
  shapes = c(0, 1, 2, 5)
)
print(g)
}
}
