% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_plot.R
\name{PlotPyramids}
\alias{PlotPyramids}
\title{Plot A Family Of Population Pyramids}
\usage{
PlotPyramids(df)
}
\arguments{
\item{df}{Long, skinny dataframe of population data}
}
\value{
Graphics grob
}
\description{
Plot A Family Of Population Pyramids
}
\examples{
\dontrun{
df <- pacehrh:::gatherPopulation(results[[1]]$Population)
g <- pacehrh::PlotPyramids(df)
print(g)
}
}
