% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_plot_fertility.R
\name{PlotResultsFertilityRates}
\alias{PlotResultsFertilityRates}
\title{Plot A Single Pair of Fertility Rates Curves From A Results List}
\usage{
PlotResultsFertilityRates(results, trial = 1, year = 2020)
}
\arguments{
\item{results}{Results list (as returned by \code{RunExperiments()})}

\item{trial}{Trail number (index into the results list)}

\item{year}{Year in trial timeseries to plot}
}
\value{
ggplot grob, or NULL on error
}
\description{
Plot A Single Pair of Fertility Rates Curves From A Results List
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)

g <- ehep::PlotResultsFertilityRates(results, 49, 2030)
print(g)
}
}
