% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_plot.R
\name{PlotResultsPopulationCurve}
\alias{PlotResultsPopulationCurve}
\title{Plot A Single Population Curve From A Results List}
\usage{
PlotResultsPopulationCurve(results, trial = 1, year = 2020, sex = "f", ...)
}
\arguments{
\item{results}{Results list (as returned by \code{\link[=RunExperiments]{RunExperiments()}})}

\item{trial}{Trial number (index into the results list)}

\item{year}{Year in trial timeseries to plot}

\item{sex}{Sex to plot (from c("m", "M", "f", "F"), defaults to female)}

\item{...}{Parameters passed to \code{\link[=PlotPopulationCurve]{PlotPopulationCurve()}}}
}
\value{
ggplot grob, or NULL on error
}
\description{
More intuitive interface to \code{\link[=PlotPopulationCurve]{PlotPopulationCurve()}}
}
\examples{
\dontrun{
# These two function calls produce the same output:
pacehrh::PlotResultsPopulationCurve(results,
  trial = 26, year = 2025, sex = "m")
pacehrh::PlotPopulationCurve(results[[26]]$Population[["2025"]]$Male)
}
}
