% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_pop_change_params_class.R
\docType{class}
\name{PopulationChangeParameters-class}
\alias{PopulationChangeParameters-class}
\alias{PopulationChangeParameters}
\title{Population Change Parameters Class}
\value{
Class of type \code{PopulationChangeParameters}
}
\description{
Population Change Parameters Class
}
\section{Slots}{

\describe{
\item{\code{values}}{numeric}
}}

