% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_pop_pyramid_class.R
\docType{class}
\name{PopulationPyramid-class}
\alias{PopulationPyramid-class}
\alias{PopulationPyramid}
\title{Population Pyramid Class}
\value{
Class of type \code{PopulationPyramid}
}
\description{
Population Pyramid Class
}
\section{Slots}{

\describe{
\item{\code{length}}{numeric}

\item{\code{values}}{numeric}
}}

