% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_experiment.R
\name{RunExperiment}
\alias{RunExperiment}
\title{Run A PACE-HRH Modeling Experiment}
\usage{
RunExperiment(debug = FALSE)
}
\arguments{
\item{debug}{(default = FALSE)}
}
\value{
List of dataframes of per-task times, or NULL
}
\description{
Apply stochasticity to the values in the BVE (\code{baseValuesEnvironment}), save the
new parameter values to the EXP \code{experimentValuesEnvironment}, then run a full
set of calculations.
}
\details{
Results are written back into \code{experimentValuesEnvironment} and as
a results structure.
}
