% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_experiments.R
\name{RunExperiments}
\alias{RunExperiments}
\title{Run A Suite Of PACE-HRH Modeling Experiments}
\usage{
RunExperiments(
  scenarioName = "ScenarioA",
  trials = 100,
  debug = FALSE,
  seed = GPE$rngSeed
)
}
\arguments{
\item{scenarioName}{(default = "ScenarioA")}

\item{trials}{(default = 100)}

\item{debug}{(default = FALSE)}

\item{seed}{Random number generator seed to be used for this suite.
(default = `GPE$rngSeed`, as set up during package initialization)}
}
\value{
List of dataframes of per-task times, or NULL
}
\description{
Results are written back into \code{experimentValuesEnvironment}.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
