% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_exp_control.R
\name{SaveBaseSettings}
\alias{SaveBaseSettings}
\title{Save Base Values For A Stochastic Experiment}
\usage{
SaveBaseSettings(scenarioName = "")
}
\arguments{
\item{scenarioName}{Scenario name string}
}
\value{
Scenario information, or NULL in case of error (such as an invalid
scenario)
}
\description{
Copy configuration parameters from the Global Package Environment to the
Base Values Environment.
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()
pacehrh::InitializeCadreRoles()

scenario <- "ScenarioName"

result <- pacehrh::SaveBaseSettings(scenario)
}
}
