% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_save_cadre_data.R
\name{SaveCadreData}
\alias{SaveCadreData}
\title{Save Cadre Information From A Suite Of Experiments}
\usage{
SaveCadreData(filepath = "cadre.csv", run = "")
}
\arguments{
\item{filepath}{CSV file to write to (default = "cadre.csv")}
}
\value{
NULL (invisible)
}
\description{
Save Cadre Information From A Suite Of Experiments
}
\examples{
\dontrun{

CHANGE THIS EXAMPLE

scenario <- "ScenarioName"

results <-
 pacehrh::RunExperiments(scenarioName = scenario,
                      trials = 100)

df <- pacehrh::SaveSuiteDemographics(results, breaks = c(50))

df <- df \%>\%
  dplyr::group_by(Trial, Year, AgeBucket) \%>\%
  dplyr::summarize(Female = sum(Female), Male = sum(Male))
}
}
