% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_save_cadre_data.R
\name{SaveCadreOverheadData}
\alias{SaveCadreOverheadData}
\title{Save Cadre Information From A Suite Of Experiments}
\usage{
SaveCadreOverheadData(filepath = NULL, run = "Run-1")
}
\arguments{
\item{filepath}{CSV file to write to (default = NULL)}

\item{run}{Identifier for the experiment run}
}
\value{
Cadre overhead data data table
}
\description{
Save Cadre Information From A Suite Of Experiments
}
\examples{
\dontrun{
}
}
