% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_setFromVector_method.R
\name{setFromVector,PopulationChangeParameters,numeric-method}
\alias{setFromVector,PopulationChangeParameters,numeric-method}
\title{Set Population Change Parameters}
\usage{
\S4method{setFromVector}{PopulationChangeParameters,numeric}(object, values)
}
\arguments{
\item{object}{PopulationChangeParameters object}

\item{values}{Vector of values to copy into PopulationChangeParameters object}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Population Change Parameters
}
