% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_setFromVector_method.R
\name{setFromVector,PopulationPyramid,numeric-method}
\alias{setFromVector,PopulationPyramid,numeric-method}
\title{Set Pyramid Age Values}
\usage{
\S4method{setFromVector}{PopulationPyramid,numeric}(object, values)
}
\arguments{
\item{object}{PopulationPyramid object}

\item{values}{Vector of values to copy into PopulationPyramid object}
}
\value{
Updated \code{PopulationPyramid} object
}
\description{
Set Pyramid Age Values
}
