% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_config.R
\name{SetInputExcelFile}
\alias{SetInputExcelFile}
\title{Set Input Excel File}
\usage{
SetInputExcelFile(inputExcelFilePath = "./config/model_inputs.xlsx")
}
\arguments{
\item{inputExcelFilePath}{Path to Excel file (default =
"./config/model_inputs.xlsx")}
}
\value{
TRUE/FALSE
}
\description{
Declare the Excel file to be used for reading configuration information. The
function over-writes whatever is supplied in the globalconfig.json file. The
function returns a warning if the user supplies a bad file path.
}
\examples{
\dontrun{
pacehrh::SetInputExcelFile(inputExcelFilePath = "config_file.xlsx")
}
}
