% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_config.R
\name{SetStochasticity}
\alias{SetStochasticity}
\title{Turn Stochasticity On/Off}
\usage{
SetStochasticity(value = NULL)
}
\arguments{
\item{value}{Desired stochasticity. (Calling \code{\link[=SetStochasticity]{SetStochasticity()}} with
no parameters or value = NULL returns the current state.)}
}
\value{
Previous stochasticity flag value (invisible)
}
\description{
Debugging function to turn off all Monte Carlo variation. With stochasticity
off, every experiment based on the same configuration produces the
same output.
By default the system is set up with stochasticity on.
}
\examples{
\dontrun{
pacehrh::SetStochasticity(FALSE) # Turn off stochasticity
pacehrh::SetStochasticity() # Return current stochasticity state
}
}
