% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_setValue_method.R
\name{setValue,PopulationChangeParameters,character,numeric-method}
\alias{setValue,PopulationChangeParameters,character,numeric-method}
\title{Set Population Change Parameter Values}
\usage{
\S4method{setValue}{PopulationChangeParameters,character,numeric}(object, label, value)
}
\arguments{
\item{object}{\code{PopulationChangeParameters} object}

\item{label}{Name of variable to write}

\item{value}{Value to write to variable}
}
\value{
Updated \code{PopulationChangeParameters} object
}
\description{
Set Population Change Parameter Values
}
