% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_task_time.R
\name{TaskTimes}
\alias{TaskTimes}
\title{Calculate Task Times Based On BVE and EXP Values}
\usage{
TaskTimes()
}
\value{
List of two matrices:
Time - annual task times in minutes
N - number of times task was performed
}
\description{
Calculate Task Times Based On BVE and EXP Values
}
