% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_trace.R
\name{Trace}
\alias{Trace}
\title{Turn Package Tracing On/Off}
\usage{
Trace(state = NULL)
}
\arguments{
\item{state}{TRUE/FALSE to set tracing state, NULL or empty to return
current state.}
}
\value{
Original state
}
\description{
Turn Package Tracing On/Off
}
\examples{
oldState <- Trace()
print(oldState)

}
