% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_task_time_stochasticity.R
\name{varyTaskValues}
\alias{varyTaskValues}
\title{Apply Stochasticity To Select Task Parameters}
\usage{
varyTaskValues(tasks)
}
\arguments{
\item{tasks}{Task parameter value matrix}
}
\value{
Adjusted task parameters matrix
}
\description{
Apply Stochasticity To Select Task Parameters
}
