% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_population_utils.R
\name{ComputeApplicablePopulationMatrices}
\alias{ComputeApplicablePopulationMatrices}
\title{Generate Matrices of Applicable Populations}
\usage{
ComputeApplicablePopulationMatrices(
  results,
  labels = BVE$populationLabels$Labels
)
}
\arguments{
\item{results}{Results structure as returned by \code{\link[=RunExperiments]{RunExperiments()}}}

\item{labels}{Population labels structure}
}
\value{
List of applicable population matrices (labels x years)
}
\description{
Generate Matrices of Applicable Populations
}
\examples{
\dontrun{
}
}
