% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_post_process.R
\name{ComputeSummaryStats}
\alias{ComputeSummaryStats}
\title{Compute Summary Statistics}
\usage{
ComputeSummaryStats(DR = NULL, CA = NULL)
}
\arguments{
\item{DR}{From [ReadAndCollateSuiteResults()] function}

\item{CA}{From [ComputeCadreAllocations()] function}
}
\value{
List of computed summary tables
}
\description{
Compute Summary Statistics
}
\examples{
\dontrun{

results <- pacehrh::RunExperiments(scenarioName = "scenario", trials = 10)
DR <- pacehrh::SaveExtendedSuiteResults(results)
CA <- pacehrh::SaveCadreAllocations(DR)
stats <- pacehrh::ComputeSummaryStats(DR, CA)
}
}
