% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_scenarios_config.R
\name{InitializeScenarios}
\alias{InitializeScenarios}
\title{Initialize Experiment Scenario Information}
\usage{
InitializeScenarios(loadFromExcel = TRUE, ...)
}
\arguments{
\item{loadFromExcel}{If TRUE, initialize the scenarios list from the model
inputs Excel file. If FALSE, initialize with a blank scenarios table.}

\item{...}{See \code{loadScenarios()}}
}
\value{
NULL (invisible)
}
\description{
Read experiment scenario information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
