% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ehep_seasonality_config.R
\name{InitializeSeasonality}
\alias{InitializeSeasonality}
\title{Initialize Seasonality Information}
\usage{
InitializeSeasonality(...)
}
\arguments{
\item{...}{Parameters passed through to \code{\link[=loadSeasonalityCurves]{loadSeasonalityCurves()}} and
\code{\link[=loadSeasonalityOffsets]{loadSeasonalityOffsets()}}}
}
\value{
NULL (invisible)
}
\description{
Read the healthcare seasonality information from the model inputs Excel file, and
save to a location in the global package environment.
}
\examples{
\dontrun{
library(ehep)

ehep::InitializePopulation()
ehep::InitializeHealthcareTasks()
ehep::InitializeScenarios()
ehep::InitializeStochasticParameters()
ehep::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  ehep::RunExperiments(scenarioName = scenario,
                       trials = 100)
}
}
