% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_seasonality_config.R
\name{loadSeasonalityOffsets}
\alias{loadSeasonalityOffsets}
\title{Load Seasonality Offsets}
\usage{
loadSeasonalityOffsets(sheetNameOffsets = .defaultSeasonalityOffsetsSheet, ...)
}
\arguments{
\item{sheetNameOffsets}{Sheet name from the model input Excel file}

\item{...}{Throw away other parameters}
}
\value{
Seasonality offsets data frame.
}
\description{
Read the seasonality offsets from the model inputs Excel file. The
name and location of the model inputs Excel file is loaded from the global
configuration JSON file.
}
