% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_initialize.R
\name{PaceInitialize}
\alias{PaceInitialize}
\title{Initialize PACE-HRH}
\usage{
PaceInitialize(
  globalConfigFile = NULL,
  inputFile = NULL,
  populationSheet = NULL,
  scenariosSheet = NULL,
  stochasticParametersSheet = NULL,
  seasonalityCurvesSheet = NULL,
  seasonalityOffsetsSheet = NULL,
  forceGlobalConfigReload = TRUE
)
}
\arguments{
\item{globalConfigFile}{Path to the global configuration JSON file}

\item{inputFile}{Path to the model inputs Excel file}

\item{populationSheet}{Baseline population sheet name}

\item{scenariosSheet}{Scenarios sheet name}

\item{stochasticParametersSheet}{Stochastic parameters sheet name}

\item{seasonalityCurvesSheet}{Seasonality curves sheet name}

\item{seasonalityOffsetsSheet}{Seasonality offsets sheet name}

\item{forceGlobalConfigReload}{TRUE/FALSE, default = TRUE. Force a reload of
the global configuration if `globalConfigFile` is specified.}
}
\value{
TRUE/FALSE
}
\description{
Initialize PACE-HRH
}
\examples{
\dontrun{
}
}
