% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_plot.R
\name{PlotMortalityRates}
\alias{PlotMortalityRates}
\title{Plot Mortality Rates}
\usage{
PlotMortalityRates(populationRates, year)
}
\arguments{
\item{populationRates}{Population rates list, as returned by
\code{RunExperiments()}}

\item{year}{Year}
}
\value{
ggplot grob
}
\description{
Plot Mortality Rates
}
\examples{
\dontrun{
library(pacehrh)

pacehrh::InitializePopulation()
pacehrh::InitializeScenarios()
pacehrh::InitializeStochasticParameters()
pacehrh::InitializeSeasonality()

scenario <- "ScenarioName"

results <-
  pacehrh::RunExperiments(scenarioName = scenario,
                       trials = 100)

g <- pacehrh::PlotMortalityRates(results[[49]]$PopulationRates, 2030)
print(g)
}
}
