% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_seasonality_experiment.R
\name{runSeasonalityExperiment}
\alias{runSeasonalityExperiment}
\title{Run The Seasonality Component Of A PACE-HRH Modeling Experiment}
\usage{
runSeasonalityExperiment(results, debug = FALSE)
}
\arguments{
\item{results}{Results from the annual part of the run}

\item{debug}{(default = FALSE)}
}
\value{
List of dataframes of per-task times, or NULL
}
\description{
Run The Seasonality Component Of A PACE-HRH Modeling Experiment
}
