% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_cadre_process.R
\name{SaveCadreAllocations}
\alias{SaveCadreAllocations}
\title{Compute And Save Cadre Allocations}
\usage{
SaveCadreAllocations(suiteResults, filepath = NULL, annual = TRUE)
}
\arguments{
\item{suiteResults}{Extended suite results as returned by `SaveExtendedSuiteResults()`}

\item{filepath}{File location to save cadre allocations. Default = NULL}

\item{annual}{Report as annual times instead of monthly? Default = TRUE}
}
\value{
data.table
}
\description{
Different healthcare delivery models involve different mixtures of types of
healthcare workers ("Cadres"). `SaveCadreAllocations()` uses cadre
information read earlier as part of running an experiment, then combines the
allocation percentages with computed task times to produce a table showing
how much time each cadre spends per year on each healthcare task.
}
\examples{
\dontrun{
resultsFiles <- c("results/results_BasicModel_Jul04.csv",
                  "results/results_ComprehensiveModel_Jul04.csv",
                  "results/results_MergedModel_Jul04.csv")

DR <- pacehrh::ReadAndCollateSuiteResults(files = resultsFiles)
CA <- pacehrh:::ComputeCadreAllocations(DR)
}
}
