% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pace_config.R
\name{SetPerAgeStats}
\alias{SetPerAgeStats}
\title{Turn Per-Age Statistics On/Off}
\usage{
SetPerAgeStats(value = NULL)
}
\arguments{
\item{value}{Desired value. Allowed values = {"none" | "monthly" | "annual"}.
Calling \code{\link[=SetPerAgeStats]{SetPerAgeStats()}} with no parameters or value = NULL returns
the current state.}
}
\value{
Previous value (invisible)
}
\description{
Turn Per-Age Statistics On/Off
}
\examples{
\dontrun{
pacehrh::SetPerAgeStats("monthly") # Turn on per-age statistics
pacehrh::SetPerAgeStats() # Return current state
}
}
